Added Manual Spawning of Infinite Tables
It is now possible to use the "up" arrow key to spawn in an infinite number of tables every three seconds. The tables all spawn in the same location, and only one furniture item can be spawned currently. This is a precursor to the ability to spawn in a dynamic amount of furniture for the room.