Added Lighting System
Added a lighting structure to the "CopyScene" scene for our usage. We could either leave it at a static height, or we could just make it a child of the ceiling as it is now. I decided to make it similar to how we are doing the floor, with a large static object that becomes more or less visible when the wall dimensions change. The structure is called "Lighting System."