Added more details, scaling options, changed room to minimum size
Added four florescent "lights" to the top of the room (the do not actually glow) and a projector to the right wall. This is to match the layout of the room in the provided picture. There are now two possible options for room scaling: by scaling the object that named "Room," one can scale the furniture and the room size all together. Meanwhile, by scaling the object inside Room called "Cube," only the walls and ceiling of the cube scale, leaving the furniture alone. Additionally, I adjusted the cube to a size that I would like to propose to be our minimum room size.