Added Manual Control over room specs in scripts
Added control over room specs such as room dimensions and furniture spawning via the scripts FurnitureList.cs and SetScale. The goal is to get these two points of change connected to the menu, but for now, we know that once we do get the menu hooked up, the mechanics should work as intended.