Added VR options menu and movement options.
Added a new menu that manages VR specific movement options. Snap turning allows the player to use the control sticks to rotate the camera a set angle to either side, and teleportation allows the player to teleport to where their controller is pointing. Teleportation is still a work in progress, but snap turning and its toggle is functional.