Bind (most) input to the Input System package
this effectively:
- binds the MouseLock and ObjectPickup scripts to be called through the PlayerController, instead of having their own update loops based on GetKeyDown checks
- moves all functionality of PauseToMain to the PlayerController, thus removing the need for the script/component. furthermore, since the script only ever seems to return to the menu, and no other scene is enabled in the build settings, it's current hardcoded to return to scene 0. this can be changed if needed, as the original code has been left commented out.

PlayerDistanceTracker was not moved over to this system, as i'm assuming that the script will eventually be changed to support automatic period output without the need for input.

as a result of these changes, the ObjectPickup component is moved to the Participant GameObject in the DynamicScene. all respective keybinds can be found in the PlayerControls.inputactions asset under the Assets\Resources directory.