Fixed Character Menu Errors
Fixed two errors in the character menu. The first error was fixing the issue that caused which character the player selected to not be saved between pauses. The second one was that when the player tried to select an option beyond the first two in the character selector, an error would be output. Now it does not spit out an error, but it does still display the wrong models since I can't create or implement them.