Changed Menu Flow and Added Furniture Room Overlay
I changed the menu flow such that the player must set their room and character options before they start the game. I also added an overlay that tells the player to use "E" to move furniture in furniture mode. It now only tells the player to record in the actual "Play" mode.