Emote wheel improvements + JSON testing
Emote button changed to being an invisible button in the middle of the screen to fit with the wheel / drag and drop setup. JSON test code added; needs to be formatted to A) actually be JSON, and B) not be an incoherent mess that just writes every frame? Shrug. Also need to test on mobile later.